+++ packages/agent-eval/src/cli/commands/scenario.ts
@@ -1,11 +1,11 @@
 import path from 'node:path'
+import Docker from 'dockerode'
 import {defineCommand} from 'citty'
 import {DefaultHost as host} from '../../host'
 import {logger} from '../../logger'
 import {
   copilotConcurrencyOption,
   containerConcurrencyOption,
-  dockerImageOption,
   getConcurrencyValue,
   getCopilotToken,
   getOutputPath,
@@ -16,6 +16,7 @@ import {
 } from '../options'
 import {getScenario} from '../../scenario/get'
 import {createScenarioPlan} from '../../scenario/plan'
+import {buildScenarioImage} from '../../scenario/scenario'
 import {runPlan} from '../../plan'
 import type {RunTrialResult} from '../../trial/run'
 
@@ -25,6 +26,109 @@ const scenarioCommand = defineCommand({
     description: 'Run scenarios',
   },
   subCommands: {
+    image: defineCommand({
+      meta: {
+        name: 'image',
+      },
+      subCommands: {
+        build: defineCommand({
+          meta: {
+            name: 'build',
+            description: 'Build the Docker image for a scenario',
+          },
+          args: {
+            name: {
+              type: 'positional',
+              description: 'The name of the scenario',
+              required: true,
+            },
+            scenarios: scenariosOption,
+          },
+          async run({args}) {
+            logger.info('Building Docker image for scenario: %s', args.name)
+
+            const scenariosDirectory = path.resolve(args.scenarios)
+            const scenario = await getScenario({
+              directory: scenariosDirectory,
+              name: args.name,
+            })
+
+            await buildScenarioImage({
+              scenario,
+            })
+          },
+        }),
+        clean: defineCommand({
+          meta: {
+            name: 'clean',
+            description: 'Remove the Docker image for a scenario, or all scenarios',
+          },
+          args: {
+            name: {
+              type: 'positional',
+              description: 'The name of the scenario',
+              required: false,
+            },
+            scenarios: scenariosOption,
+          },
+          async run({args}) {
+            const scenariosDirectory = path.resolve(args.scenarios)
+            const docker = new Docker()
+
+            if (args.name) {
+              logger.info('Removing Docker images for scenario: %s', args.name)
+              const images = await docker.listImages().then(images => {
+                return images.filter(image => {
+                  return image.RepoTags?.some(tag => {
+                    return tag.startsWith(`agent-eval/scenarios/${args.name}:`)
+                  })
+                })
+              })
+
+              for (const image of images) {
+                if (image.RepoTags === undefined) {
+                  throw new Error(`Image ${image.Id} has no tags`)
+                }
+
+                const [tag] = image.RepoTags
+
+                logger.info('Removing image with tag: %s', tag)
+                const dockerImage = docker.getImage(tag)
+                await dockerImage.remove({
+                  force: true,
+                })
+              }
+            } else {
+              logger.info('Removing all Docker images for scenarios in directory: %s', scenariosDirectory)
+              const tagGroups = ['agent-eval/tools', 'agent-eval/sandbox', 'agent-eval/scenarios']
+              const images = await docker.listImages().then(images => {
+                return images.filter(image => {
+                  return image.RepoTags?.some(tag => {
+                    return tagGroups.some(group => {
+                      return tag.startsWith(group)
+                    })
+                  })
+                })
+              })
+
+              for (const image of images) {
+                if (image.RepoTags === undefined) {
+                  throw new Error(`Image ${image.Id} has no tags`)
+                }
+
+                const [tag] = image.RepoTags
+
+                logger.info('Removing image with tag: %s', tag)
+                const dockerImage = docker.getImage(tag)
+                await dockerImage.remove({
+                  force: true,
+                })
+              }
+            }
+          },
+        }),
+      },
+    }),
     run: defineCommand({
       meta: {
         name: 'run',
@@ -37,7 +141,6 @@ const scenarioCommand = defineCommand({
         },
         'copilot-concurrency': copilotConcurrencyOption,
         'container-concurrency': containerConcurrencyOption,
-        'docker-image': dockerImageOption,
         name: {
           type: 'positional',
           description: 'The name of the scenario',
@@ -96,7 +199,6 @@ const scenarioCommand = defineCommand({
           copilotConcurrency,
           containerConcurrency,
           copilotToken,
-          dockerImage: args['docker-image'],
           plan,
         })